# Some perf related changes when profiling the MI Task example within spiff-arena