--- a/SpiffWorkflow/bpmn/util/task.py
+++ b/SpiffWorkflow/bpmn/util/task.py
@@ -32,11 +32,16 @@ class BpmnTaskFilter(TaskFilter):
         def _catches_event(task):
             return isinstance(task.task_spec, CatchingEvent) and task.task_spec.catches(task, self.catches_event)
 
-        return all([
-            super().matches(task),
-            self.catches_event is None or _catches_event(task),
-            self.lane is None or task.task_spec.lane == self.lane,
-        ])
+        if not super().matches(task):
+            return False
+
+        if not (self.catches_event is None or _catches_event(task)):
+            return False
+
+        if not (self.lane is None or task.task_spec.lane == self.lane):
+            return False
+
+        return True
 
 
 class BpmnTaskIterator(TaskIterator):
@@ -55,12 +60,11 @@ class BpmnTaskIterator(TaskIterator):
         task = self.task_list.pop(0)
         subprocess = task.workflow.top_workflow.subprocesses.get(task.id)
 
-        if all([
-            len(task._children) > 0 or subprocess is not None,
-            task.state >= self.min_state or subprocess is not None,
-            self.depth < self.max_depth,
-            task.task_spec.name != self.end_at_spec,
-        ]):
+        if (len(task._children) > 0 or subprocess is not None) and \
+            (task.state >= self.min_state or subprocess is not None) and \
+            self.depth < self.max_depth and \
+            task.task_spec.name != self.end_at_spec:
+            
             # Do not descend into a completed subprocess to look for unfinished tasks.
             if (
                 subprocess is None