--- a/SpiffWorkflow/util/task.py
+++ b/SpiffWorkflow/util/task.py
@@ -225,7 +225,7 @@ class TaskIterator:
 
     def _next(self):
 
-        if len(self.task_list) == 0:
+        if not self.task_list:
             raise StopIteration()
 
         task = self.task_list.pop(0)
@@ -239,7 +239,7 @@ class TaskIterator:
             else:
                 self.task_list.extend(task.children)
             self._update_depth(task)
-        elif self.depth_first and len(self.task_list) > 0:
+        elif self.depth_first and self.task_list:
             self._handle_leaf_depth(task)
     
         return task